Rename clients utility.hpp to clients_utility.hpp
There is a clients/rocblascommon/utility.hpp and a
library/rocblascommon/utility.hpp and the identical names
cause confusion.

In particular, common_host_helpers.hpp includes utility.hpp, which
results in including a different file when the host helpers are used
in the library vs. when the host helpers are used in the clients.
