+++ clients/include/clientcommon.hpp
@@ -10,7 +10,7 @@
 // rocblas common
 #include "rocblascommon/rocblas_test.hpp"
 #include "rocblascommon/rocblas_vector.hpp"
-#include "rocblascommon/utility.hpp"
+#include "rocblascommon/clients_utility.hpp"
 //#include "rocblascommon/device_vector.hpp"
 //#include "rocblascommon/device_batch_vector.hpp"
 //#include "rocblascommon/device_strided_batch_vector.hpp"
